+++ jabgui/src/main/java/org/jabref/gui/fieldeditors/LinkedFilesEditor.java
@@ -309,6 +309,9 @@ public class LinkedFilesEditor extends VBox implements FieldEditorFX {
 
         HBox container = new HBox(2);
         container.setPrefHeight(Double.NEGATIVE_INFINITY);
+        container.getStyleClass().add("file-row");
+        EasyBind.subscribe(linkedFile.isAutomaticallyFoundProperty(),
+                found -> container.pseudoClassStateChanged(PseudoClass.getPseudoClass("auto-found"), found));
         container.getChildren().addAll(acceptAutoLinkedFile, info, writeMetadataToPdf, parsePdfMetadata);
 
         return container;